pyproject.toml
[tool.poetry]
name = "opencode-on-im"
version = "0.1.0"
description = "Connect OpenCode AI coding assistant to IM platforms (Telegram, DingTalk)"
authors = ["OpenCode Cloudify Team"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/opencode-cloudify/opencode-on-im"
repository = "https://github.com/opencode-cloudify/opencode-on-im"
keywords = ["opencode", "telegram", "dingtalk", "ai", "coding-assistant", "bot"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
packages = [{include = "opencode_on_im", from = "src"}]

[tool.poetry.scripts]
opencode-on-im = "opencode_on_im.__main__:main"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.4"
dingtalk-stream = "^0.24"
pydantic = "^2.6"
pydantic-settings = "^2.2"
aiohttp = "^3.9"
aiofiles = "^23.2"
pillow = "^10.2"
pygments = "^2.17"
qrcode = {extras = ["pil"], version = "^7.4"}
aiosqlite = "^0.19"
structlog = "^24.1"
tenacity = "^8.2"
click = "^8.1"
rich = "^13.7"
httpx = "^0.27"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.23"
pytest-cov = "^4.1"
ruff = "^0.2"
mypy = "^1.8"
pre-commit = "^3.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = ["E501"]  # line length handled by formatter

[tool.ruff.lint.isort]
known-first-party = ["opencode_on_im"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=opencode_on_im --cov-report=term-missing"

[tool.coverage.run]
source = ["src/opencode_on_im"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
